inflearn/hello-spring/hello-spring/src/main/java/HELLO/hellospring/repository/MemoryMemberRepository.java
package HELLO.hellospring.repository;

import HELLO.hellospring.domain.Member;

import javax.swing.text.html.Option;
import java.util.*;

public class MemoryMemberRepository implements MemberRepository {

    private static Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;

    @Override
    public Member save(Member member){
        member.setId(++sequence);
        store.put(member.getId(),member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        // optional은 null이 반환될 경우를 대비하여 사용
        return Optional.ofNullable(store.get(id));
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream() // 반복
                .filter(member -> member.getName().equals(name)) // parameter로 넘어온 name과 같은 것 필터링
                .findAny(); // 찾은 것 아무거나 반환
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore(){
        store.clear();
    }
}
